commenting the script

+++ scripts_Bacteria/5.join_reads.sh
@@ -4,14 +4,15 @@ datapath="/home/ngs/200206_M04028_0114_000000000-CWB58/"
 outdir="Analysis/bacteria"
 core=8
 
-if [ ! -d "Analysis/${outdir}/micca_16S" ]; then
-	mkdir Analysis/${outdir}/micca_16S
+if [ ! -d "${outdir}/micca_16S" ]; then
+	mkdir -p ${outdir}/micca_16S
 fi
 
 
 # Count trimmed data
 cd $currpath/${outdir}/trimmed
 
+echo " - counting sequences "
 for i in *.fastq.gz
 do
         echo -n $i >> seq_count_16S_QC.txt
@@ -26,18 +27,22 @@ cd $currpath/${outdir}/trimmed
 
 # remove singles reads from sickle
 
-rm ./*singles.gz
+if [ -f ./*singles.gz ]; then
+	rm ./*singles.gz
+fi
 
+echo " - uncompressing trimmed fastq files "
 gunzip *.fastq.gz
 
 ## SINGULARITY CONTAINER ##
+echo " - joining reads"
 singularity run $currpath/micca.sif micca mergepairs -i $currpath/${outdir}/trimmed/*_R1.fastq -o $currpath/${outdir}/micca_16S/WP1_assembled_16S.fastq -l 32 -d 8 -t 7
 
 # -l : minimum overlap between reads
 # -d : maximum mismatch in overlap region
 
 # Counting reads in assembled file
-
+echo " - counting reads after joining "
 grep -c '^@M' $currpath/${outdir}/micca_16S/WP1_assembled_16S.fastq
 
 # 11534421. Total sum of trimmed reads = 26670364; Teoric 100% assembly = 26670364/2 = 13335182
@@ -45,7 +50,7 @@ grep -c '^@M' $currpath/${outdir}/micca_16S/WP1_assembled_16S.fastq
 # Read loss from QC reads to assembled % = 1800761/13335182*100 = 13%
 
 # zipping back trimmed files
-
+echo " - recompressing trimmed fastq files "
 cd $currpath/${outdir}/trimmed
 
 gzip *.fastq
